+++ Tone/instrument/SimpleAM.js
@@ -46,13 +46,12 @@ function(Tone){
 		 *  the carrier.
 		 *  @type {Positive}
 		 *  @signal
-		 *  @private
 		 *  @example
 		 * //set the modulator an octave above the carrier frequency
 		 * simpleAM.harmonicity.value = 2;
 		 */
-		this._harmonicity = new Tone.Multiply(options.harmonicity);
-		this._harmonicity.units = Tone.Type.Positive;
+		this.harmonicity = new Tone.Multiply(options.harmonicity);
+		this.harmonicity.units = Tone.Type.Positive;
 
 		/**
 		 *  convert the -1,1 output to 0,1
@@ -70,7 +69,7 @@ function(Tone){
 
 		//control the two voices frequency
 		this.frequency.connect(this.carrier.frequency);
-		this.frequency.chain(this._harmonicity, this.modulator.frequency);
+		this.frequency.chain(this.harmonicity, this.modulator.frequency);
 		this.modulator.chain(this._modulationScale, this._modulationNode.gain);
 		this.carrier.chain(this._modulationNode, this.output);
 		this._readOnly(["carrier", "modulator", "frequency"]);
@@ -141,21 +140,6 @@ function(Tone){
 	};
 
 	/**
-	 * The ratio between the two carrier and the modulator. 
-	 * @memberOf Tone.SimpleAM#
-	 * @type {number}
-	 * @name harmonicity
-	 */
-	Object.defineProperty(Tone.SimpleAM.prototype, "harmonicity", {
-		get : function(){
-			return this._harmonicity.value;
-		},
-		set : function(harm){
-			this._harmonicity.value = harm;
-		}
-	});
-
-	/**
 	 *  clean up
 	 *  @returns {Tone.SimpleAM} this
 	 */
@@ -168,8 +152,8 @@ function(Tone){
 		this.modulator = null;
 		this.frequency.dispose();
 		this.frequency = null;
-		this._harmonicity.dispose();
-		this._harmonicity = null;
+		this.harmonicity.dispose();
+		this.harmonicity = null;
 		this._modulationScale.dispose();
 		this._modulationScale = null;
 		this._modulationNode.disconnect();